+++ packages/plugins/EVM/src/state/Connection/interceptors/ContractWallet.ts
@@ -15,9 +15,9 @@ import {
 import { Web3 } from '@masknet/web3-providers'
 import WalletABI from '@masknet/web3-contracts/abis/Wallet.json'
 import type { Wallet as WalletContract } from '@masknet/web3-contracts/types/Wallet.js'
-import { RegisteredProviders } from '../provider.js'
-import type { BaseContractWalletProvider } from '../providers/BaseContractWallet.js'
 import { SharedContextSettings } from '../../../settings/index.js'
+import { RegisteredProviders } from '../../Providers/provider.js'
+import type { BaseContractWalletProvider } from '../../Providers/BaseContractWallet.js'
 
 export class ContractWallet implements Middleware<ConnectionContext> {
     constructor(
@@ -112,12 +112,12 @@ export class ContractWallet implements Middleware<ConnectionContext> {
     }
 
     async fn(context: ConnectionContext, next: () => Promise<void>) {
-        const provider = RegisteredProviders[context.providerType] as BaseContractWalletProvider | undefined
-
         if (!context.writeable) {
             await next()
             return
         }
+
+        const provider = RegisteredProviders[context.providerType] as BaseContractWalletProvider | undefined
 
         // not a SC wallet provider
         if (!provider?.ownerAccount && !context.owner) {
