+++ packages/plugins/EVM/src/state/Providers/provider.ts
@@ -1,16 +1,16 @@
 import { ProviderType } from '@masknet/web3-shared-evm'
-import type { BaseProvider } from './providers/Base.js'
-import { CustomNetworkProvider } from './providers/CustomNetwork.js'
-import { MaskWalletProvider } from './providers/MaskWallet.js'
-import { MetaMaskProvider } from './providers/MetaMask.js'
-import WalletConnectProvider from './providers/WalletConnect.js'
-import { WalletLinkProvider } from './providers/WalletLink.js'
-import { Coin98Provider } from './providers/Coin98.js'
-import { MathWalletProvider } from './providers/MathWallet.js'
-import { CloverProvider } from './providers/Clover.js'
-import FortmaticProvider from './providers/Fortmatic.js'
-import { OperaProvider } from './providers/Opera.js'
-import { NoneProvider } from './providers/None.js'
+import type { BaseProvider } from './Base.js'
+import { CustomNetworkProvider } from './CustomNetwork.js'
+import { MaskWalletProvider } from './MaskWallet.js'
+import { MetaMaskProvider } from './MetaMask.js'
+import WalletConnectProvider from './WalletConnect.js'
+import { WalletLinkProvider } from './WalletLink.js'
+import { Coin98Provider } from './Coin98.js'
+import { MathWalletProvider } from './MathWallet.js'
+import { CloverProvider } from './Clover.js'
+import FortmaticProvider from './Fortmatic.js'
+import { OperaProvider } from './Opera.js'
+import { NoneProvider } from './None.js'
 
 /**
  * Register all supported providers
